requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
black==23.12.0
certifi==2022.6.15
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cloudinary==1.37.0
crispy-bootstrap5==0.6
cryptography==41.0.5
Django==4.2.7
django-autoslug==1.9.8
django-colorfield==0.7.2
django-countries==7.3.2
django-crispy-forms==1.14.0
djangorestframework==3.14.0
ecdsa==0.18.0
frozenlist==1.4.1
idna==3.3
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==9.2.0
pinax-eventlog==5.1.1
platformdirs==4.1.0
psycopg2-binary==2.9.3
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.2
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
validators==0.22.0
yarl==1.9.4
